+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - [PR #2315](https://github.com/rqlite/rqlite/pull/2315): Include "before" and "after" in CDC events.
 - [PR #2318](https://github.com/rqlite/rqlite/pull/2318): Add ColumnNames() to DB object.
 - [PR #2319](https://github.com/rqlite/rqlite/pull/2319): CDC "before" and "after" now structured as maps.
+- [PR #2321](https://github.com/rqlite/rqlite/pull/2321): Add metrics to CDC Service.
 
 ## v8.43.4 (August 27th 2025)
 ### Implementation changes and bug fixes